feat(server.ts): added ssl capability
