+++ scripts/create-ssl.sh
@@ -1,0 +1,5 @@
+#!/bin/zsh
+
+mkdir -p ssl
+openssl genrsa -out ssl/server.key 2048
+openssl req -new -x509 -key ssl/server.key -out ssl/server.cert -days 365
