+++ src/modules.d.ts
@@ -6,13 +6,3 @@ declare module "express" {
         user?: JWTUser;
     }
 }
-
-declare namespace NodeJS {
-    export interface ProcessEnv {
-        NODE_ENV: "development" | "production";
-        ACCESS_TOKEN_SECRET: string;
-        SESSION_SECRET: string;
-        MONGODB_URI_LOCAL: string;
-        MONGODB_URI: string;
-    }
-}
